Add a new channel type for muxed clients.
Previously the slave client process did nothing but wait for an
exit message from the master. Now, it handles all keyboard input
and forwards the data to the master. This input is performed in
the new SSH_CHANNEL_MUX_OPEN channel type in the slave process.
The master still runs a regular SSH_CHANNEL_OPEN session to the
remote host on behalf of the slave, and it still writes all of
this channel output directly to the slave tty, but it no longer
does any reading, and the channels rfd is set to -1.

Splitting things apart like this allows each session to operate
their ttys independently without mixing up mode changes. It also
allows the slave client to suspend itself on demand.

Further additions to the mux framework include a WINCH message;
the slave client handles its own SIGWINCH events and sends the
data to the master, which forwards it to the remote host. This
eliminates the need for the master to poll every session with
TIOCGWINSZ and send size updates for every session at once.

Also, to support readline, the master session has an output filter
to search for a commandline prompt on the tail of the incoming
data. This also gets sent from the master to the slave process.

Likewise, tty-change messages received for the session are forwarded
from the master to the slave process.